+++ dashboard/src/components/Row/index.tsx
@@ -1,0 +1,6 @@
+// Copyright 2020-2023 the Kubeapps contributors.
+// SPDX-License-Identifier: Apache-2.0
+
+import { Row } from "./Row";
+
+export default Row;
