--- dashboard/src/components/js/Column/index.js
@@ -1,5 +1,0 @@
-// Copyright 2020-2022 the Kubeapps contributors.
-// SPDX-License-Identifier: Apache-2.0
-
-import Column from "./Column";
-export default Column;
