--- dashboard/src/components/js/Row/index.js
@@ -1,5 +1,0 @@
-// Copyright 2020-2022 the Kubeapps contributors.
-// SPDX-License-Identifier: Apache-2.0
-
-import Row from "./Row";
-export default Row;
